refactor: mover entidades e enums para domain/model
