+++ src/main/java/com/sensedia/sample/consents/domain/model/ConsentHistory.java
@@ -1,5 +1,6 @@
-package com.sensedia.sample.consents.domain;
+package com.sensedia.sample.consents.domain.model;
 
+import com.sensedia.sample.consents.domain.enums.ConsentStatus;
 import io.swagger.v3.oas.annotations.media.Schema;
 import lombok.*;
 import org.springframework.data.annotation.Id;
