+++ src/main/java/com/sensedia/sample/consents/dto/ConsentRequestDTO.java
@@ -1,6 +1,6 @@
 package com.sensedia.sample.consents.dto;
 
-import com.sensedia.sample.consents.domain.ConsentStatus;
+import com.sensedia.sample.consents.domain.enums.ConsentStatus;
 import io.swagger.v3.oas.annotations.media.Schema;
 import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.NotNull;
