+++ src/main/java/com/sensedia/sample/consents/repository/ConsentHistoryRepository.java
@@ -1,6 +1,6 @@
 package com.sensedia.sample.consents.repository;
 
-import com.sensedia.sample.consents.domain.ConsentHistory;
+import com.sensedia.sample.consents.domain.model.ConsentHistory;
 import org.springframework.data.mongodb.repository.MongoRepository;
 import org.springframework.stereotype.Repository;
