+++ src/test/java/com/sensedia/sample/consents/integration/ConsentApiIntegrationTest.java
@@ -1,8 +1,8 @@
 package com.sensedia.sample.consents.integration;
 
 import com.sensedia.sample.consents.ConsentsApplication;
-import com.sensedia.sample.consents.domain.Consent;
-import com.sensedia.sample.consents.domain.ConsentStatus;
+import com.sensedia.sample.consents.domain.model.Consent;
+import com.sensedia.sample.consents.domain.enums.ConsentStatus;
 import com.sensedia.sample.consents.dto.ConsentRequestDTO;
 import com.sensedia.sample.consents.dto.ConsentUpdateDTO;
 import com.sensedia.sample.consents.repository.ConsentRepository;
